+++ crates/vertigo/src/tests/dom/mod.rs
@@ -1,4 +1,5 @@
 mod html_attrs;
 mod component;
+mod component_dynamic;
 mod list_spread;
 mod params;